Add ASCII art representation of GNU/Linux
This commit introduces a new ASCII art file, classic.txt, featuring a visual representation of the GNU/Linux logo. The addition enhances the aesthetic appeal of the project and provides users with a fun way to display system information. The ASCII art is designed to be easily integrated into the Fletch tool for a more engaging user experience.